Lookups tests added to engines app.
